Add SettingsViewDelegate and UserIterations
