+++ IOS4_HW12_MIKHAIL_BAGMET/Sources/Scenes/Settings/Controller/Protocols/SettingsViewDelegate.swift
@@ -1,0 +1,12 @@
+//
+//  SettingsViewDelegate.swift
+//  IOS4_HW12_MIKHAIL_BAGMET
+//
+//  Created by Михаил on 30.01.2022.
+//
+
+import Foundation
+
+protocol SettingsViewDelegate {
+    func changeViewController(cellName: String)
+}
