+++ IOS4_HW12_MIKHAIL_BAGMET/Sources/Scenes/Settings/View/Protocols/SettingsViewUserIterations.swift
@@ -1,0 +1,12 @@
+//
+//  SettingsViewUserIterations.swift
+//  IOS4_HW12_MIKHAIL_BAGMET
+//
+//  Created by Михаил on 31.01.2022.
+//
+
+import Foundation
+
+protocol SettingsViewUserIterations {
+    var delegate: SettingsViewDelegate? { get set }
+}
